requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
supabase==1.0.3
requests==2.31.0
